Extrapolation

Extrapolation of splines. See here for some examples.

Functions

BSplineKit.SplineExtrapolations.extrapolate — Function

extrapolate(S::Union{Spline, SplineWrapper}, method::AbstractExtrapolationMethod)

Construct a SplineExtrapolation from the given spline S (which can also be the result of an interpolation).

Extrapolation methods

BSplineKit.SplineExtrapolations.Flat — Type

Flat <: AbstractExtrapolationMethod

Represents a flat extrapolation: spline values at domain limits are extended to the left and to the right.

BSplineKit.SplineExtrapolations.Smooth — Type

Smooth <: AbstractExtrapolationMethod

Represents a smooth extrapolation: derivatives up to order $k - 2$ are continuous at the boundaries.

Internals

BSplineKit.SplineExtrapolations.AbstractExtrapolationMethod — Type

AbstractExtrapolationMethod

Abstract type representing an extrapolation method.

BSplineKit.SplineExtrapolations.SplineExtrapolation — Type

SplineExtrapolation

Represents a spline which can be evaluated outside of its limits according to a given extrapolation method.