Splines

Splines

BSplineKit.Splines.Spline — Type

Spline{T}

Represents a spline function.

Spline(B::AbstractBSplineBasis, coefs::AbstractVector)

Construct a spline from a B-spline basis and a vector of B-spline coefficients.

Examples

julia> B = BSplineBasis(BSplineOrder(4), -1:0.2:1);

julia> coefs = rand(length(B));

julia> S = Spline(B, coefs)
13-element Spline{Float64}:
 order: 4
 knots: [-1.0, -1.0, -1.0, -1.0, -0.8, -0.6, -0.4, -0.2, 0.0, 0.2, 0.4, 0.6, 0.8, 1.0, 1.0, 1.0, 1.0]
 coefficients: [0.484368, 0.973024, 0.474683, 0.77864, 0.0663495, 0.818427, 0.629158, 0.00961182, 0.180137, 0.100973, 0.444828, 0.422792, 0.438264]

Spline{T = Float64}(undef, B::AbstractBSplineBasis)

Construct a spline with uninitialised vector of coefficients.

(S::Spline)(x)

Evaluate spline at coordinate x.

BSplineKit.Splines.coefficients — Function

coefficients(S::Spline)

Get B-spline coefficients of the spline.

Base.eltype — Method

eltype(::Type{<:Spline})
eltype(S::Spline)

Returns type of element returned when evaluating the Spline.

Base.length — Method

length(S::Spline)

Returns the number of coefficients in the spline.

Note that this is equal to the number of basis functions, length(basis(S)).

BSplineKit.BSplines.basis — Method

basis(S::Spline) -> AbstractBSplineBasis

Returns the associated B-spline basis.

Derivatives and integrals

Base.:* — Function

*(op::Derivative, S::Spline) -> Spline

Returns N-th derivative of spline S as a new spline.

See also diff.

Examples

julia> B = BSplineBasis(BSplineOrder(4), -1:0.2:1);

julia> S = Spline(B, rand(length(B)))
13-element Spline{Float64}:
 order: 4
 knots: [-1.0, -1.0, -1.0, -1.0, -0.8, -0.6, -0.4, -0.2, 0.0, 0.2, 0.4, 0.6, 0.8, 1.0, 1.0, 1.0, 1.0]
 coefficients: [0.411779, 0.477218, 0.602547, 0.30413, 0.304017, 0.750247, 0.614036, 0.519966, 0.657764, 0.437818, 0.107356, 0.429024, 0.435861]

julia> Derivative(1) * S
12-element Spline{Float64}:
 order: 3
 knots: [-1.0, -1.0, -1.0, -0.8, -0.6, -0.4, -0.2, 0.0, 0.2, 0.4, 0.6, 0.8, 1.0, 1.0, 1.0]
 coefficients: [0.981589, 0.939965, -1.49208, -0.000564486, 2.23115, -0.681057, -0.470348, 0.688991, -1.09973, -1.65231, 2.41251, 0.102554]

julia> Derivative(2) * S
11-element Spline{Float64}:
 order: 2
 knots: [-1.0, -1.0, -0.8, -0.6, -0.4, -0.2, 0.0, 0.2, 0.4, 0.6, 0.8, 1.0, 1.0]
 coefficients: [-0.416237, -12.1602, 7.45759, 11.1586, -14.561, 1.05354, 5.7967, -8.94361, -2.7629, 20.3241, -23.0995]

Base.diff — Function

diff(S::Spline, [op::Derivative = Derivative(1)]) -> Spline

Same as op * S.

Returns N-th derivative of spline S as a new spline.

BSplineKit.Splines.integral — Function

integral(S::Spline)

Returns an antiderivative of the given spline as a new spline.

The algorithm is described in de Boor 2001, p. 127.

Spline wrappers

BSplineKit.Splines.SplineWrapper — Type

SplineWrapper{S <: Spline}

Abstract type representing a type that wraps a Spline.

Such a type implements all common operations on splines, including evaluation, differentiation, etc…

BSplineKit.Splines.spline — Function

spline(w::SplineWrapper) -> Spline

Returns the Spline wrapped by the object.