B-splines

BSplineKit.BSplines — Module

BSplines

Module defining B-spline bases and B-spline related functions.

B-spline bases

BSplineKit.BSplines.AbstractBSplineBasis — Type

AbstractBSplineBasis{k,T}

Abstract type defining a B-spline basis, or more generally, a functional basis defined from B-splines.

The basis is represented by a B-spline order k and a knot element type T.

BSplineKit.BSplines.BSplineBasis — Type

BSplineBasis{k}

B-spline basis for splines of order k.

The basis is defined by a set of knots and by the B-spline order.

BSplineBasis(
    k::Union{BSplineOrder,Integer}, knots::Vector;
    augment::Val = Val(true),
)

Create B-spline basis of order k with the given knots.

If augment = Val(true) (default), knots will be "augmented" so that both knot ends have multiplicity k. See also augment_knots.

BSplineKit.BSplines.boundaries — Function

boundaries(B::AbstractBSplineBasis)

Returns (xmin, xmax) tuple with the boundaries of the domain supported by the basis.

BSplineKit.BSplines.order — Function

order(::Type{BSplineBasis}) -> Int
order(::Type{Spline}) -> Int
order(::BSplineOrder) -> Int

Returns order of B-splines as an integer.

BSplineKit.BSplines.knots — Function

knots(g::BSplineBasis)
knots(g::Spline)

Returns the knots of the B-spline basis.

Base.getindex — Function

getindex(B::AbstractBSplineBasis, i, [T = Float64])

Get $i$-th basis function.

This is an alias for BasisFunction(B, i, T) (see BasisFunction for details).

The returned object can be evaluated at any point within the boundaries defined by the basis.

Examples

julia> B = BSplineBasis(BSplineOrder(4), -1:0.1:1)
23-element BSplineBasis: order 4, domain [-1.0, 1.0]

julia> B[6]
Basis function i = 6
  from 23-element BSplineBasis: order 4, domain [-1.0, 1.0]
  support: [-0.8, -0.4) (knots 6:10)

julia> B[6](-0.5)
0.16666666666666666

julia> B[6, Float32](-0.5)
0.16666667f0

julia> B[6](-0.5, Derivative(1))
-5.000000000000001

Base.length — Method

length(g::BSplineBasis)

Returns the number of B-splines composing a spline.

Basis functions

BSplineKit.BSplines.BasisFunction — Type

BasisFunction{B <: AbstractBSplineBasis, T}

Describes a single basis function.

The basis function may belong to a BSplineBasis (in which case it's effectively a B-spline), or to a basis derived from a B-spline basis (such as a RecombinedBSplineBasis).

BasisFunction(basis::AbstractBSplineBasis, i::Int, [T = Float64])

Construct i-th basis function of the given basis.

The constructed function can be evaluated as b(x), returning a value of type T.

(b::BasisFunction)(x, [op::AbstractDifferentialOp])

Evaluate basis function at coordinate x.

To evaluate a derivative, pass Derivative(n) as the op argument, with n the derivative order.

More general differential operators, such as Derivative(n) + λ Derivative(m), are also supported.

BSplineKit.BSplines.support — Function

support(b::BasisFunction) -> UnitRange{Int}

Get range of knots supported by the basis function.

Returns the knot range i:j such that the basis function support is $t ∈ [t_i, t_j)$.

support(B::BSplineBasis, i::Integer) -> UnitRange{Int}

Get range of knots supported by the $i$-th basis function.

BSplineKit.BSplines.common_support — Function

common_support(b1::BasisFunction, b2::BasisFunction, ...) -> UnitRange{Int}

Get range of knots commonly supported by different basis functions.

If the supports don't intersect, an empty range is returned (e.g. 6:5), following the behaviour of intersect. The lack of intersection can be checked using isempty, which returns true for such a range.

BSplineKit.BSplines.evaluate — Function

evaluate(B::AbstractBSplineBasis, i::Integer, x,
         [op::AbstractDifferentialOp], [T=Float64])

Evaluate $i$-th basis function in the given basis at x (can be a coordinate or a vector of coordinates).

To evaluate a derivative, pass Derivative(n) as the op argument, with n the derivative order.

More general differential operators, such as Derivative(n) + λ Derivative(m), are also supported.

See also evaluate!.

BSplineKit.BSplines.evaluate! — Function

evaluate!(b::AbstractVector, B::BSplineBasis, i::Integer,
          x::AbstractVector, args...)

Evaluate i-th basis function at positions x and write result to b.

See also evaluate.

Internals

BSplineKit.BSplines.BSplineOrder — Type

BSplineOrder(k::Integer)

Specifies the B-spline order k.

BSplineKit.BSplines.augment_knots — Function

augment_knots(knots::AbstractVector, k::Union{Integer,BSplineOrder})

Modifies the input knots to make sure that the first and last knot have multiplicity k for splines of order k.

It is assumed that border knots have multiplicity 1 at the borders. That is, border coordinates should not be repeated in the input.

BSplineKit.BSplines.multiplicity — Function

multiplicity(knots, i)

Determine multiplicity of knot knots[i].