Interpolation

High-level interpolation of gridded data and function approximation using B-splines.

Functions

Interpolations.interpolate — Function

interpolate(x, y, BSplineOrder(k))

Interpolate values y at locations x using B-splines of order k.

Grid points x must be real-valued and are assumed to be in increasing order.

Returns a SplineInterpolation which can be evaluated at any intermediate point.

See also interpolate!.

Examples

julia> xs = -1:0.1:1;

julia> ys = cospi.(xs);

julia> itp = interpolate(xs, ys, BSplineOrder(4));

julia> itp(-1)
-1.0

julia> itp(0)
1.0

julia> itp(0.42)
0.2486897676885842

Interpolations.interpolate! — Function

interpolate!(I::SplineInterpolation, y::AbstractVector)

Update spline interpolation with new data.

This function allows to reuse a SplineInterpolation returned by a previous call to interpolate, using new data on the same locations x.

See interpolate for details.

BSplineKit.SplineInterpolations.approximate — Function

approximate(f, B::AbstractBSplineBasis, [x = collocation_points(B)]) -> SplineInterpolation

Approximate function f in the given basis.

The approximation is performed by interpolation of a discrete set of evaluated values $y_i = f(x_i)$, where the data points $x_i$ may be given as input.

The returned SplineInterpolation approximates the given function.

Example

julia> B = BSplineBasis(BSplineOrder(3), -1:0.4:1);

julia> S = approximate(sin, B)
SplineInterpolation containing the 7-element Spline{Float64}:
 order: 3
 knots: [-1.0, -1.0, -1.0, -0.6, -0.2, 0.2, 0.6, 1.0, 1.0, 1.0]
 coefficients: [-0.8414709848078965, -0.7317273726556252, -0.39726989430226317, 0.0, 0.39726989430226317, 0.7317273726556252, 0.8414709848078965]
 interpolation points: [-1.0, -0.8, -0.4, 0.0, 0.4, 0.8, 1.0]

BSplineKit.SplineInterpolations.approximate! — Function

approximate!(f, S::SplineInterpolation)

Approximate function f in the basis associated to the SplineInterpolation S.

See also approximate.

BSplineKit.SplineInterpolations.spline — Function

spline(I::SplineInterpolation) -> Spline

Returns the Spline associated to the interpolation.

Types

BSplineKit.SplineInterpolations.SplineInterpolation — Type

SplineInterpolation

Spline interpolation.

This is the type returned by interpolate.

A SplineInterpolation I can be evaluated at any point x using the I(x) syntax.

It can also be updated with new data on the same data points using interpolate!.

SplineInterpolation(undef, B::AbstractBSplineBasis, x::AbstractVector, [T = eltype(x)])

Initialise a SplineInterpolation from B-spline basis and a set of interpolation (or collocation) points x.

Note that the length of x must be equal to the number of B-splines.

Use interpolate! to actually interpolate data known on the x locations.