Interpolations

High-level interpolation of gridded data using B-splines.

Interpolation functions

BSplineKit.Interpolations.interpolate — Function

interpolate(x, y, k::Integer)
interpolate(x, y, BSplineOrder(k))

Interpolate values y at locations x using B-splines of order k.

Grid points x must be real-valued and are assumed to be in increasing order.

Returns an Interpolation which can be evaluated at any intermediate point.

The second form using BSplineOrder ensures that the type of the returned Interpolation is fully inferred by the compiler.

See also interpolate!.

BSplineKit.Interpolations.interpolate! — Function

interpolate!(I::Interpolation, y::AbstractVector)

Update spline interpolation with new data.

This function allows to reuse an Interpolation returned by a previous call to interpolate, using new data on the same locations x.

See interpolate for details.

Interpolation objects

BSplineKit.Interpolations.Interpolation — Type

Interpolation

Spline interpolation.

This is the type returned by interpolate. Generally, it should not be directly constructed.

An Interpolation I can be evaluated at any point x using the I(x) syntax.

It can also be updated with new data on the same data points using interpolate!.

BSplineKit.Interpolations.spline — Function

spline(I::Interpolation) -> Spline

Returns the Spline associated to the interpolation.