Splines

Splines

BSplineKit.Splines.Spline — Type

Spline

Spline function.

Spline(b::BSplineBasis, coefs::AbstractVector)

Construct a spline from a B-spline basis and a vector of B-spline coefficients.

Spline(undef, B::BSplineBasis, [T=Float64])

Construct a spline with uninitialised vector of coefficients.

(S::Spline)(x)

Evaluate spline at coordinate x.

BSplineKit.Splines.coefficients — Function

coefficients(S::Spline)

Get B-spline coefficients of the spline.

Base.eltype — Method

eltype(S::Spline)

Returns type of element returned when evaluating the Spline.

Base.length — Method

length(S::Spline)

Returns the number of coefficients in the spline.

Note that this is equal to the number of basis functions, length(basis(S)).

BSplineKit.BSplines.basis — Method

basis(S::Spline) -> BSplineBasis

Returns the associated B-spline basis.

Derivatives and integrals

Base.diff — Function

diff(S::Spline, [deriv::Derivative = Derivative(1)]) -> Spline

Return N-th derivative of spline S as a new spline.

BSplineKit.Splines.integral — Function

integral(S::Spline)

Returns an antiderivative of the given spline as a new spline.

The algorithm is described in de Boor 2001, p. 127.