Splines

Splines

BSplineKit.Splines.Spline — Type

Spline{N, T}

Represents an $N$-dimensional spline function which returns values of type T.

For $N ≥ 2$, this is a tensor-product spline. That is, the spline space is given by the tensor product of $N$ one-dimensional spline spaces.

Spline(B::AbstractBSplineBasis, coefs::AbstractVector)

Construct a 1D spline from a B-spline basis and a vector of B-spline coefficients.

The spline can be then evaluated at any point x within the domain.

Examples

1D splines

julia> B = BSplineBasis(BSplineOrder(4), -1:0.2:1);

julia> coefs = rand(length(B));

julia> S = Spline(B, coefs)
13-element Spline{1, Float64}:
 basis: 13-element BSplineBasis of order 4, domain [-1.0, 1.0]
 order: 4
 knots: [-1.0, -1.0, -1.0, -1.0, -0.8, -0.6, -0.4, -0.2, 0.0, 0.2, 0.4, 0.6, 0.8, 1.0, 1.0, 1.0, 1.0]
 coefficients: [0.173575, 0.321662, 0.258585, 0.166439, 0.527015, 0.483022, 0.390663, 0.802763, 0.721983, 0.372347, 0.0301856, 0.0793339, 0.663758]

julia> S(0.42)  # evaluate spline
0.6543543311366747

julia> S′ = Derivative(1) * S  # spline derivative
12-element Spline{1, Float64}:
 basis: 12-element BSplineBasis of order 3, domain [-1.0, 1.0]
 order: 3
 knots: [-1.0, -1.0, -1.0, -0.8, -0.6, -0.4, -0.2, 0.0, 0.2, 0.4, 0.6, 0.8, 1.0, 1.0, 1.0]
 coefficients: [2.22131, -0.473071, -0.460734, 1.80288, -0.219964, -0.461794, 2.0605, -0.403899, -1.74818, -1.71081, 0.368613, 8.76636]

julia> Sint = integral(S)  # spline integral
14-element Spline{1, Float64}:
 basis: 14-element BSplineBasis of order 5, domain [-1.0, 1.0]
 order: 5
 knots: [-1.0, -1.0, -1.0, -1.0, -1.0, -0.8, -0.6, -0.4, -0.2, 0.0, 0.2, 0.4, 0.6, 0.8, 1.0, 1.0, 1.0, 1.0, 1.0]
 coefficients: [0.0, 0.00867873, 0.0408449, 0.0796327, 0.11292, 0.218323, 0.314928, 0.393061, 0.553613, 0.69801, 0.772479, 0.777007, 0.78494, 0.818128]

Multidimensional (tensor-product) splines

julia> Bx = BSplineBasis(BSplineOrder(4), -1:0.2:1)
13-element BSplineBasis of order 4, domain [-1.0, 1.0]
 knots: [-1.0, -1.0, -1.0, -1.0, -0.8, -0.6, -0.4, -0.2, 0.0, 0.2, 0.4, 0.6, 0.8, 1.0, 1.0, 1.0, 1.0]

julia> By = BSplineBasis(BSplineOrder(6), 0:0.05:1)
25-element BSplineBasis of order 6, domain [0.0, 1.0]
 knots: [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.05, 0.1, 0.15, 0.2  …  0.8, 0.85, 0.9, 0.95, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0]

julia> coefs = rand(length(Bx), length(By));

julia> S = Spline((Bx, By), coefs)
13×25 Spline{2, Float64}:
 bases:
   (1) 13-element BSplineBasis of order 4, domain [-1.0, 1.0]
   (2) 25-element BSplineBasis of order 6, domain [0.0, 1.0]
 knots:
   (1) [-1.0, -1.0, -1.0, -1.0, -0.8, -0.6, -0.4, -0.2, 0.0, 0.2, 0.4, 0.6, 0.8, 1.0, 1.0, 1.0, 1.0]
   (2) [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.05, 0.1, 0.15, 0.2  …  0.8, 0.85, 0.9, 0.95, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0]
 coefficients: [0.173575 0.183429 … 0.00993008 0.320248; 0.321662 0.347508 … 0.251083 0.346025; … ; 0.433914 0.895838 … 0.483176 0.0793339; 0.211228 0.580212 … 0.425628 0.663758]

julia> S(-0.32, 0.54)
0.38520723534067336

Spline(undef, B::AbstractBSplineBasis, [T = Float64])
Spline(undef, (Bx, By, …), [T = Float64])

Construct a spline with uninitialised vector of coefficients.

In the second case, a tensor-product multidimensional spline is constructed.

The optional parameter T corresponds to the returned type when the spline is evaluated.

BSplineKit.Splines.Spline1D — Type

Spline1D

Alias for Spline{1}, representing a one-dimensional spline.

BSplineKit.Splines.coefficients — Function

coefficients(S::Spline{N,T}) -> AbstractArray{T,N}

Returns the array of B-spline coefficients of the spline.

Base.eltype — Method

eltype(::Type{<:Spline})
eltype(S::Spline)

Returns type of element returned when evaluating the Spline.

Base.length — Method

length(S::Spline)

Returns the number of coefficients in the spline.

For 1D splines, this is equal to the number of basis functions, length(first(bases(S))).

Base.size — Method

size(S::Spline)

Same as size(coefficients(S)).

Base.ndims — Method

ndims(::Type{<:Spline})
ndims(S::Spline)

Returns the dimensionality of the spline.

BSplineKit.Splines.bases — Method

bases(S::Spline) -> (B₁, B₂, …)

Returns the B-spline bases associated to the spline.

The number of bases is equal to the dimensionality of the spline. In particular, for 1D splines, this returns a one-element tuple.

See also basis(::Spline1D).

BSplineKit.BSplines.basis — Method

basis(S::Spline1D)

Returns the B-spline basis associated to the spline.

Derivatives and integrals

Base.:* — Function

*(op::Derivative, S::Spline1D) -> Spline1D

Returns N-th derivative of spline S as a new spline.

See also diff.

Examples

julia> B = BSplineBasis(BSplineOrder(4), -1:0.2:1);

julia> S = Spline(B, rand(length(B)))
13-element Spline{1, Float64}:
 basis: 13-element BSplineBasis of order 4, domain [-1.0, 1.0]
 order: 4
 knots: [-1.0, -1.0, -1.0, -1.0, -0.8, -0.6, -0.4, -0.2, 0.0, 0.2, 0.4, 0.6, 0.8, 1.0, 1.0, 1.0, 1.0]
 coefficients: [0.173575, 0.321662, 0.258585, 0.166439, 0.527015, 0.483022, 0.390663, 0.802763, 0.721983, 0.372347, 0.0301856, 0.0793339, 0.663758]

julia> Derivative(0) * S === S
true

julia> Derivative(1) * S
12-element Spline{1, Float64}:
 basis: 12-element BSplineBasis of order 3, domain [-1.0, 1.0]
 order: 3
 knots: [-1.0, -1.0, -1.0, -0.8, -0.6, -0.4, -0.2, 0.0, 0.2, 0.4, 0.6, 0.8, 1.0, 1.0, 1.0]
 coefficients: [2.22131, -0.473071, -0.460734, 1.80288, -0.219964, -0.461794, 2.0605, -0.403899, -1.74818, -1.71081, 0.368613, 8.76636]

julia> Derivative(2) * S
11-element Spline{1, Float64}:
 basis: 11-element BSplineBasis of order 2, domain [-1.0, 1.0]
 order: 2
 knots: [-1.0, -1.0, -0.8, -0.6, -0.4, -0.2, 0.0, 0.2, 0.4, 0.6, 0.8, 1.0, 1.0]
 coefficients: [-26.9438, 0.0616849, 11.3181, -10.1142, -1.20915, 12.6114, -12.322, -6.72141, 0.186876, 10.3971, 83.9775]

Base.diff — Function

diff(S::Spline1D, [op::Derivative = Derivative(1)]) -> Spline1D

Same as op * S.

Returns N-th derivative of spline S as a new spline.

BSplineKit.Splines.integral — Function

integral(S::Spline1D)

Returns an antiderivative of the given spline as a new spline.

The algorithm is described in de Boor 2001, p. 127.

Spline wrappers

BSplineKit.Splines.SplineWrapper — Type

SplineWrapper{S <: Spline}

Abstract type representing a type that wraps a Spline.

Such a type implements all common operations on splines, including evaluation, differentiation, etc…

BSplineKit.Splines.spline — Function

spline(w::SplineWrapper) -> Spline

Returns the Spline wrapped by the object.